fix: osc client compat (countdowns)

+++ src/protocol/osc/client.cpp
@@ -60,7 +60,7 @@ struct param_visitor : public boost::static_visitor<void>
     }
 
     void operator()(const bool value) { o << value; }
-    void operator()(const int32_t value) { o << static_cast<int64_t>(value); }
+    void operator()(const int32_t value) { o << static_cast<int32_t>(value); }
     void operator()(const uint32_t value) { o << static_cast<int64_t>(value); }
     void operator()(const int64_t value) { o << static_cast<int64_t>(value); }
     void operator()(const uint64_t value) { o << static_cast<int64_t>(value); }
